Fixes after merge with main branch:
- move guideMenu container boxes inside the component (use client directive is needed)
- fix getTutorial and getGuide api functions
- remove pages folder
- upgrade editorial-components dependency
