+++ apps/nextjs-website/package.json
@@ -20,7 +20,7 @@
     "@mui/utils": "^5.14.5",
     "@mui/material": "^5.14.5",
     "@pagopa/mui-italia": "1.0.1",
-    "@pagopa/pagopa-editorial-components": "2.0.0",
+    "@pagopa/pagopa-editorial-components": "2.1.3",
     "@stoplight/elements": "^7.9.0",
     "fp-ts": "^2.13.1",
     "gitbook-docs": "*",
