+++ apps/nextjs-website/src/app/[productSlug]/guides/[...productGuidePage]/page.tsx
@@ -1,17 +1,19 @@
 import Layout, { LayoutProps } from '@/components/organisms/Layout/Layout';
-import { getGuide, getGuidePaths } from '@/lib/api';
+import { getGuide, getGuidePaths, getProducts } from '@/lib/api';
 import { Product } from '@/lib/types/product';
 import { Box } from '@mui/material';
 import React from 'react';
+import GuideMenu from '@/components/atoms/GuideMenu/GuideMenu';
 import GitBookContent from '@/components/organisms/GitBookContent/GitBookContent';
-import EContainer from '@pagopa/pagopa-editorial-components/dist/components/EContainer';
-import GitbookLeftMenu from '@/components/molecules/GitbookLeftMenu/GitbookLeftMenu';
+import GuideInPageMenu from '@/components/organisms/GuideInPageMenu/GuideInPageMenu';
+import { FragmentProvider } from '@/components/organisms/FragmentProvider/FragmentProvider';
+import { gitBookPagesWithTitle, spaceToPrefixMap } from '@/_contents/products';
+import { translations } from '@/_contents/translations';
+import { ParseContentConfig } from 'gitbook-docs/parseContent';
 
-type GuideParams = {
-  params: {
-    productSlug: string;
-    productGuidePage: Array<string>;
-  };
+type Params = {
+  productSlug: string;
+  productGuidePage: Array<string>;
 };
 
 export async function generateStaticParams() {
@@ -34,17 +36,38 @@ type ProductGuidePageProps = {
   }[];
   path: string;
   pathPrefix: string;
-  assetsPrefix: string;
   isIndex: boolean;
   menu: string;
   body: string;
+  bodyConfig: ParseContentConfig;
 } & LayoutProps;
 
-const Page = async ({ params }: GuideParams) => {
-  const props: ProductGuidePageProps = await getGuide(
-    params.productSlug,
-    params.productGuidePage
-  );
+const Page = async (params: Params) => {
+  const { productGuidePage } = translations;
+
+  const productSlug = params?.productSlug;
+  const guidePath = params?.productGuidePage.join('/');
+  const path = `/${productSlug}/guides/${guidePath}`;
+  const guideProps = await getGuide(path);
+  const { product, page, guide, version, versions, source, bannerLinks } =
+    guideProps;
+  const props: ProductGuidePageProps = {
+    ...page,
+    product,
+    products: [...getProducts()],
+    guide,
+    version,
+    versions,
+    bannerLinks,
+    pathPrefix: source.pathPrefix,
+    bodyConfig: {
+      isPageIndex: page.isIndex,
+      pagePath: page.path,
+      assetsPrefix: source.assetsPrefix,
+      gitBookPagesWithTitle,
+      spaceToPrefix: spaceToPrefixMap,
+    },
+  };
 
   return (
     <Layout
@@ -54,31 +77,58 @@ const Page = async ({ params }: GuideParams) => {
       bannerLinks={props.bannerLinks}
       showBreadcrumbs={false}
     >
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: { xs: 'column-reverse', lg: 'row' },
-        }}
-      >
-        <GitbookLeftMenu
-          guideName={props.guide.name}
-          versionName={props.version.name}
-          versions={props.versions}
-          menu={props.menu}
-          pathPrefix={props.pathPrefix}
-          assetsPrefix={props.assetsPrefix}
-        />
-        <EContainer>
-          <Box>
-            <GitBookContent
-              assetsPrefix={props.assetsPrefix}
-              pagePath={props.path}
-              isPageIndex={props.isIndex}
-              content={props.body}
-            />
+      <FragmentProvider>
+        <Box
+          sx={{
+            display: 'flex',
+            flexDirection: { xs: 'column-reverse', lg: 'row' },
+            margin: '0 auto',
+            maxWidth: '1900px',
+          }}
+        >
+          <GuideMenu
+            menu={props.menu}
+            assetsPrefix={props.bodyConfig.assetsPrefix}
+            linkPrefix={props.pathPrefix}
+            guideName={props.guide.name}
+            versionName={props.version.name}
+            versions={props.versions}
+          />
+          <Box
+            sx={{
+              margin: '0 auto',
+              maxWidth: '1008px',
+              padding: '56px 40px',
+              flexGrow: { lg: 1 },
+            }}
+          >
+            <GitBookContent content={props.body} config={props.bodyConfig} />
           </Box>
-        </EContainer>
-      </Box>
+          <Box
+            sx={{
+              display: { xs: 'none', lg: 'initial' },
+              position: 'relative',
+              padding: { lg: '80px 64px' },
+              width: { lg: '270px' },
+            }}
+          >
+            <Box
+              sx={{
+                position: 'sticky',
+                maxWidth: '270px',
+                top: 20,
+              }}
+            >
+              <GuideInPageMenu
+                assetsPrefix={props.bodyConfig.assetsPrefix}
+                pagePath={props.path}
+                inPageMenu={props.body}
+                title={productGuidePage.onThisPage}
+              />
+            </Box>
+          </Box>
+        </Box>
+      </FragmentProvider>
     </Layout>
   );
 };
