+++ apps/nextjs-website/src/app/[productSlug]/tutorials/[...productTutorialPage]/page.tsx
@@ -1,15 +1,15 @@
 import Layout, { LayoutProps } from '@/components/organisms/Layout/Layout';
 import EContainer from '@pagopa/pagopa-editorial-components/dist/components/EContainer';
-import { getTutorial, getTutorialPaths } from '@/lib/api';
+import { getTutorial, getTutorialPaths, getProducts } from '@/lib/api';
 import { Product } from '@/lib/types/product';
 import GitBookContent from '@/components/organisms/GitBookContent/GitBookContent';
 import { Box } from '@mui/material';
+import { gitBookPagesWithTitle, spaceToPrefixMap } from '@/_contents/products';
+import { ParseContentConfig } from 'gitbook-docs/parseContent';
 
-type TutorialParams = {
-  params: {
-    productSlug: string;
-    productTutorialPage: Array<string>;
-  };
+type Params = {
+  productSlug: string;
+  productTutorialPage: Array<string>;
 };
 
 export async function generateStaticParams() {
@@ -22,16 +22,29 @@ export async function generateStaticParams() {
 type ProductTutorialPageProps = {
   product: Product;
   path: string;
-  pathPrefix: string;
-  assetsPrefix: string;
   body: string;
+  bodyConfig: ParseContentConfig;
 } & LayoutProps;
 
-const Page = async ({ params }: TutorialParams) => {
-  const props: ProductTutorialPageProps = await getTutorial(
-    params.productSlug,
-    params.productTutorialPage
-  );
+const Page = async (params: Params) => {
+  const productSlug = params?.productSlug;
+  const tutorialPath = params?.productTutorialPage.join('/');
+  const path = `/${productSlug}/tutorials/${tutorialPath}`;
+  const tutorialProps = await getTutorial(path);
+  const { product, page, bannerLinks, source } = tutorialProps;
+  const props: ProductTutorialPageProps = {
+    ...page,
+    product,
+    products: [...getProducts()],
+    bannerLinks,
+    bodyConfig: {
+      isPageIndex: false,
+      pagePath: page.path,
+      assetsPrefix: source.assetsPrefix,
+      gitBookPagesWithTitle,
+      spaceToPrefix: spaceToPrefixMap,
+    },
+  };
 
   return (
     <Layout
@@ -42,13 +55,8 @@ const Page = async ({ params }: TutorialParams) => {
       showBreadcrumbs={false}
     >
       <EContainer>
-        <Box>
-          <GitBookContent
-            assetsPrefix={props.assetsPrefix}
-            pagePath={props.path}
-            isPageIndex={false}
-            content={props.body}
-          />
+        <Box sx={{ padding: '56px 0' }}>
+          <GitBookContent content={props.body} config={props.bodyConfig} />
         </Box>
       </EContainer>
     </Layout>
