--- apps/nextjs-website/src/pages/[productSlug]/tutorials/[...productTutorialPage].tsx
@@ -1,77 +1,0 @@
-import Layout, { LayoutProps } from '@/components/organisms/Layout/Layout';
-import EContainer from '@pagopa/pagopa-editorial-components/dist/components/EContainer';
-import { getTutorial, getTutorialPaths, getProducts } from '@/lib/api';
-import { Product } from '@/lib/types/product';
-import GitBookContent from '@/components/organisms/GitBookContent/GitBookContent';
-import { GetStaticPaths, GetStaticProps } from 'next/types';
-import { Box } from '@mui/material';
-import { gitBookPagesWithTitle, spaceToPrefixMap } from '@/_contents/products';
-import { ParseContentConfig } from 'gitbook-docs/parseContent';
-
-type Params = {
-  productSlug: string;
-  productTutorialPage: Array<string>;
-};
-
-export const getStaticPaths: GetStaticPaths<Params> = () => {
-  return {
-    paths: [...getTutorialPaths()],
-    fallback: false,
-  };
-};
-
-type ProductTutorialPageProps = {
-  product: Product;
-  path: string;
-  body: string;
-  bodyConfig: ParseContentConfig;
-} & LayoutProps;
-
-export const getStaticProps: GetStaticProps<
-  ProductTutorialPageProps,
-  Params
-> = ({ params }) => {
-  const productSlug = params?.productSlug;
-  const tutorialPath = params?.productTutorialPage.join('/');
-  const path = `/${productSlug}/tutorials/${tutorialPath}`;
-  const tutorialProps = getTutorial(path);
-  if (tutorialProps) {
-    const { product, page, bannerLinks, source } = tutorialProps;
-    const props = {
-      ...page,
-      product,
-      products: [...getProducts()],
-      bannerLinks,
-      bodyConfig: {
-        isPageIndex: false,
-        pagePath: page.path,
-        assetsPrefix: source.assetsPrefix,
-        gitBookPagesWithTitle,
-        spaceToPrefix: spaceToPrefixMap,
-      },
-    };
-    return { props };
-  } else {
-    return { notFound: true as const };
-  }
-};
-
-const Page = (props: ProductTutorialPageProps) => {
-  return (
-    <Layout
-      products={props.products}
-      product={props.product}
-      path={props.path}
-      bannerLinks={props.bannerLinks}
-      showBreadcrumbs={false}
-    >
-      <EContainer>
-        <Box sx={{ padding: '56px 0' }}>
-          <GitBookContent content={props.body} config={props.bodyConfig} />
-        </Box>
-      </EContainer>
-    </Layout>
-  );
-};
-
-export default Page;
